config,lfsapi: move and export httpconfig type
