+++ config/httpconfig.go
@@ -1,4 +1,4 @@
-package lfsapi
+package config
 
 import (
 	"fmt"
@@ -6,15 +6,21 @@ import (
 	"strings"
 )
 
-type httpconfig struct {
-	git Env
+type HTTPConfig struct {
+	git Environment
+}
+
+func NewHTTPConfig(git Environment) *HTTPConfig {
+	return &HTTPConfig{
+		git: git,
+	}
 }
 
 // Get retrieves a `http.{url}.{key}` for the given key and urls, following the
 // rules in https://git-scm.com/docs/git-config#git-config-httplturlgt.
 // The value for `http.{key}` is returned as a fallback if no config keys are
 // set for the given urls.
-func (c *httpconfig) Get(key string, rawurl string) (string, bool) {
+func (c *HTTPConfig) Get(key string, rawurl string) (string, bool) {
 	key = strings.ToLower(key)
 	if v, ok := c.get(key, rawurl); ok {
 		return v, ok
@@ -22,7 +28,7 @@ func (c *httpconfig) Get(key string, rawurl string) (string, bool) {
 	return c.git.Get(fmt.Sprintf("http.%s", key))
 }
 
-func (c *httpconfig) get(key, rawurl string) (string, bool) {
+func (c *HTTPConfig) get(key, rawurl string) (string, bool) {
 	u, err := url.Parse(rawurl)
 	if err != nil {
 		return "", false
@@ -37,14 +43,14 @@ func (c *httpconfig) get(key, rawurl string) (string, bool) {
 	pLen := len(u.Path)
 	if pLen > 2 {
 		end := pLen
-		if strings.HasSuffix(u.Path, slash) {
+		if strings.HasSuffix(u.Path, "/") {
 			end -= 1
 		}
 
-		paths := strings.Split(u.Path[1:end], slash)
+		paths := strings.Split(u.Path[1:end], "/")
 		for i := len(paths); i > 0; i-- {
 			for _, host := range hosts {
-				path := strings.Join(paths[:i], slash)
+				path := strings.Join(paths[:i], "/")
 				if v, ok := c.git.Get(fmt.Sprintf("http.%s/%s.%s", host, path, key)); ok {
 					return v, ok
 				}
